[#2] Move Signature from stamp to view
Not sure if it's going to stay this way, but there was no need for it
inside the stamp.

Also tweaked BDocView's constructor to require a \DOMDocument. A static
method was added to create views from signer & files - the raw
constructor is expected to be used when creating a view from an already
existing signature file.